+++ src/finestvx/localization/__init__.py
@@ -1,8 +1,0 @@
-"""Localization exports for FinestVX."""
-
-from .service import LocalizationConfig, create_localization
-
-__all__ = [
-    "LocalizationConfig",
-    "create_localization",
-]